Little work on training

+++ LoaderPACK/Loader.py
@@ -81,6 +81,8 @@ class load_shuffle_5_min(Dataset):
             chan = int(i//cuts_pr_chan) # the given channel
             inp = self.ls[0][0][chan][cut_point[i]:cut_point[i]+60*5*250].view(1, 60*5*250)
             tar = self.ls[1][0][chan][cut_point[i]:cut_point[i]+60*5*250].view(1, 60*5*250)
+            #inp = self.ls[0][0][chan][cut_point[i]:cut_point[i]+60*5*250]
+            #tar = self.ls[1][0][chan][cut_point[i]:cut_point[i]+60*5*250]
             yield (inp, tar, chan)
 
     def __len__(self):
@@ -89,7 +91,7 @@ class load_shuffle_5_min(Dataset):
     def __getitem__(self, idx):
         inp, tar, chan = next(self.gen)
         inp = inp.to(self.device)
-        tar = tar.to(self.device)
+        tar = tar.to(self.device).type(torch.LongTensor)
         return inp, tar, chan
